+++ R/crypto_prices.R
@@ -6,18 +6,13 @@
 #' @return Will provide data frame of current prices
 #' @details Updated every 5 minutes
 #' @examples {
-#' kin_price <- getPrices("kin")
+#' kin_price <- crypto_prices("kin")
 #' }
-#' @rdname getPrices
+#' @rdname crypto_prices
 #' @export
 #' @importFrom jsonlite read_json
-#' @importFrom magrittr %>%
 #' @import tidyr
-getPrices <- function(coin = NULL, limit = 0, currency = NULL) {
-  ifelse(as.character(match.call()[[1]]) == "getPrices",
-    warning("DEPRECATED: Please use crypto_prices() instead of getPrices().", call. = FALSE, immediate. = TRUE),
-    shh <- ""
-    )
+crypto_prices <- function(coin = NULL, limit = 0, currency = NULL) {
   options(scipen = 999)
   url <- "https://api.coinmarketcap.com/v1/ticker/"
   if (is.null(coin)) {
@@ -36,22 +31,18 @@ getPrices <- function(coin = NULL, limit = 0, currency = NULL) {
       url <- paste0(url, "?convert=", currency)
     }
   }
-  prices <- use_rate_limit(jsonlite::read_json(url, simplifyVector = TRUE))
-  prices$market_cap_usd <- prices$market_cap_usd %>% tidyr::replace_na(0) %>% as.numeric()
+  prices                  <- jsonlite::read_json(url, simplifyVector = TRUE)
+  prices$market_cap_usd   <- prices$market_cap_usd %>% tidyr::replace_na(0) %>% as.numeric()
   prices$available_supply <- prices$available_supply %>% tidyr::replace_na(0) %>% as.numeric()
-  prices$max_supply <- prices$max_supply %>% tidyr::replace_na(0) %>% as.numeric()
-  prices$total_supply <- prices$total_supply %>% tidyr::replace_na(0) %>% as.numeric()
+  prices$max_supply       <- prices$max_supply %>% tidyr::replace_na(0) %>% as.numeric()
+  prices$total_supply     <- prices$total_supply %>% tidyr::replace_na(0) %>% as.numeric()
   cols <- c(4:14)
   prices[, cols] <- apply(prices[, cols], 2, function(x) replace(x, is.na(x), 0))
   prices[, cols] <- suppressWarnings(apply(prices[, cols], 2, function(x) as.numeric(x)))
-  prices[, 15] <- as.POSIXct(as.numeric(prices[, 15]), origin = "1970-01-01")
+  prices[, 15]   <- as.POSIXct(as.numeric(prices[, 15]), origin = "1970-01-01")
   if (!is.null(currency)) {
     concols <- c(16:18)
     prices[, concols] <- suppressWarnings(apply(prices[, concols], 2, function(x) as.numeric(x)))
   }
   return(prices)
 }
-
-#' @export
-#' @rdname getPrices
-crypto_prices <- getPrices